Initial commit with IPython Notebook and CSV
Adds the CSV file containing Density data collected over
the last year and a half, though excluding this odd hole
for about a month.

Also includes an IPython notebook that provides some functions
for displaying graphs for floors/buildings on intervals of 15
minutes, as well as weekly/daily counts.

Data is currently truncated to between 6/30/2014 and 4/19/2015
due to the hole, as well as some trickiness in handling overlap
of weeks over multiple years ("weekofyear" can often count days
at the beginning/end of a year as part of the previous/next year,
depending on how much of that week is contained which year).

TODO: incorporate overlap