Added window selection capability to plot functions
floor_plot and bldg_plot now have the ability to select for
the start-date/end-date for data to be displayed via keyword
arguments.

When using group="week", however, the start-date and end-date
are still hardcoded at 06/30/2014 and 04/19/2015; still need to
get around the weeks issue.

Also added a few more plots displaying the seasonality of
capacity throughout the week and on a daily basis.

	modified:   density.ipynb